Move composite generator to separate package
Make it clearer that these are only used in the interop tests
